+++ app/(pages)/feed/message/page.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { FC } from 'react';
-import { MessageContent } from '@/app/components';
+
 const Message: FC = () => {
-  return <MessageContent />;
+  return <div className='mt-4 space-y-4 md:pt-40'>content</div>;
 };
 
 export default Message;
