+++ app/components/FeedComponents/Sidebar/Sidebar.scss
