--- app/components/FeedComponents/index.ts
@@ -1,7 +1,0 @@
-import MessageContent from './message';
-import Search from './sectiions/search/search';
-import Sidebar from './sectiions/sidebar/sidebar';
-import { FeedPost } from './dummydata';
-import SingleFeedPost from './SingleFeedPost';
-
-export { MessageContent, Search, Sidebar, FeedPost,SingleFeedPost };
