--- app/components/FeedComponents/message.tsx
@@ -1,7 +1,0 @@
-import { FC } from 'react';
-
-const MessageContent: FC = () => {
-  return <div className='mt-4 space-y-4 md:pt-40'>content</div>;
-};
-
-export default MessageContent;
