+++ app/components/Header/Header.scss
@@ -1,9 +1,7 @@
-$prefix: header;
+$prefix: header-container;
 
 .#{$prefix} {
-  // write your CSS here
-}
-
-.vertical-spacing {
-  min-height: 4.25rem;
+  .vertical-spacing {
+    min-height: 4.25rem;
+  }
 }
